Update Ionic Modals, Forms, Popovers Assignment 2
Assignment Overview

In this assignment, you will support a popover in the dishdetail.html
template with two options. The first option will enable you to add the
dish to your favorites. The second option will bring up a form within a
modal that enables you to add your comments on the dish.

Assignment Requirements

This assignment consists of three tasks:

Task 1

Add a button to the header bar that when clicked will bring up a popover
The popover should have two options: "Add to Favorites", and "Add
Comment"
The popover template should be defined in a file named
dish-detail-popover.html
Appropriate code is added to DishDetailController to support the
popover.
Note: Download style.css file provided above and replace the style.css
file within the www/css folder within your project. This file contains
some CSS customization for the Popover.

Task 2

When the user clicks on the "Add to Favorites" option in the popover,
then add the dish to the list of your favorite dishes.
Task 3

When the user clicks on the "Add Comment" option in the popover, bring
up a modal containing the form that enables the user to add comments on
the dish
The form should contain three fields: Rating, Name and Comment. Use a
Select element for the Rating, and a textarea for Comment fields.
The form should be defined a template named dish-comment.html
The modal is properly configured within DishDetailController
When the user submits the comment, it should be added to the list of
comments for the dish.
Review criterialess
Your submission will be graded based on the following criteria:

Task 1

The nav-button has been added to the header bar
The popover is appropriately designed to show the two options in the
template file.
The popover is configured with the appropriate code in
DishDetailController
Task 2

The dish is added correctly to the list of favorites when the user
selects the option
Task 3

The modal containing the form is shown when the user selects the "Add
Comment" option.
The form is properly configured within the template
The modal is properly configured within DishDetailController
The comment is properly added to the list of comments for the dish when
the user submits the comment.