Update Cordova plugins and ngCordova wrappers of these plugins
Task 1

The registration page should now include a new button named Gallery
below the Take Picture button.
Clicking on the Gallery button should bring up the image gallery from
which the user should be able to select a picture to include in the
registration page
The appropriate ngCordova wrapper service is properly injected in the
AppCtrl controller
Task 2

When the user clicks on the Add To Favorites option in the Popover in
the dish details page, a notification should be added to the
notification bar with the information of the dish
When the user clicks on the Add To Favorites option in the Popover in
the dish details page, a toast message is shown at the bottom of the
screen showing the details of the dish added to the favorites
The appropriate ngCordova wrapper service is properly injected in the
DishDetailController controller.
Task 3

The appropriate ngCordova wrapper service is properly injected in the
FavoritesController controller.
The phone is vibrated when the user confirms that the item should be
deleted.